add some more integration tests

+++ src/main/java/io/github/jhipster/online/web/rest/AccountResource.java
@@ -159,7 +159,6 @@ public class AccountResource {
     @Timed
     public void deleteAccount() {
         final String userLogin = SecurityUtils.getCurrentUserLogin().orElseThrow(() -> new InternalServerErrorException("Current user login not found"));
-
         // Checks if user exists
         Optional<User> user = userRepository.findOneByLogin(userLogin);
         if (!user.isPresent()) {
@@ -170,8 +169,12 @@ public class AccountResource {
             jdlService.deleteAllForJdlMetadata(jdlMetadata.getId());
         }
         jdlMetadataService.deleteAllForCurrentUser(userLogin);
-        githubService.deleteAllOrganizationsForCurrentUser(userLogin);
-        gitlabService.deleteAllOrganizationsForCurrentUser(userLogin);
+        if (githubService != null) {
+            githubService.deleteAllOrganizationsForCurrentUser(userLogin);
+        }
+        if (gitlabService != null) {
+            gitlabService.deleteAllOrganizationsForCurrentUser(userLogin);
+        }
         userService.deleteUser(userLogin);
     }
